PatchManager USI Kolonization Systems
An opt-in patch using PatchManager for USI Kolonization Systems that modifies the bulkhead profile of some 'Karibou', 'MKS', and 'WOLF' parts to a better size. Also, allows for the 'Ranger' Battery and Thermal Control System parts to appear in the Kolonization category.

Adds cylindrical bulkhead tags for size12 [15m], size16 [20m], and size24 [30m]. Adds hexadecimal colors for reference.