PatchManager ChromaWorks
An opt-in patch using PatchManager for ChromaWorks that creates an alternative VAB Organizer configuration focusing on integrating the parts into subcategories based upon their purpose. In pursuit of that, it switches the categories of a few parts and replaces the bulkhead profile to support a 'PCB' bulkhead tag.

Adds a [PCB] (CircuitBoard) bulkhead tag.