Organized Readme
Combined link to VABOrganizer with name and stated it as a requirement. Revamped 'Supported Mods' section through making subheadings, ordering mods alphanumerically, consistency in the "Support from", and adding separate subheadings for common/useful subcategories and bulkhead tags. Removed number from dev branch, so it doesn't need to be updated each time. Added "the dev branch of" to submitting a pull request part.